docs(README): update usage instructions for parse and generate modes
This commit updates the README to clarify the usage instructions for
the parse and generate modes. It specifies that the `_redirects` file
should be parsed from `/public` (or following the `publicDir` config)
and that a `_redirects` file can be generated to `/public` (or following
the `publicDir` config) from the `entries` option.